+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 
 - `title`:
   - Reformat titles such as **"Title" by Artist** to **Artist - Title**.
+  - Do not move remixes to the end of the title.
 
 ## [0.20.0] 2024-11-04